Fix: Cast height to int in filament-pinpoint config

+++ config/filament-pinpoint.php
@@ -30,6 +30,6 @@ return [
         'lat' => (float) env('GOOGLE_MAPS_DEFAULT_LAT', -0.5050),
         'lng' => (float) env('GOOGLE_MAPS_DEFAULT_LNG', 117.1500),
         'zoom' => (int) env('GOOGLE_MAPS_DEFAULT_ZOOM', 13),
-        'height' => env('GOOGLE_MAPS_DEFAULT_HEIGHT', 400),
+        'height' => (int) env('GOOGLE_MAPS_DEFAULT_HEIGHT', 400),
     ],
 ];
